Functionality split up between DzrpDezogIfRemote and ZxNextSerialRemote
